The renames in template resolution classes - making names shorter - making names more objective
