+++ src/Microsoft.TemplateEngine.Cli/TemplateResolution/TemplateListResolutionResult.cs
@@ -1,244 +1,190 @@
 // Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Microsoft.TemplateEngine.Edge.Template;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Microsoft.TemplateEngine.Edge.Template;
+using Microsoft.TemplateEngine.Utils;
+using System.Collections.ObjectModel;
+using Microsoft.TemplateEngine.Abstractions;
 
 namespace Microsoft.TemplateEngine.Cli.TemplateResolution
 {
-    public class TemplateListResolutionResult
+    /// <summary>
+    /// The class represents the template resolution result for listing purposes.
+    /// The resolution result is the list of templates to use
+    /// </summary>
+    public sealed class TemplateListResolutionResult
     {
-        public TemplateListResolutionResult(string templateName, string userInputLanguage, IReadOnlyCollection<ITemplateMatchInfo> coreMatchedTemplates, IReadOnlyCollection<ITemplateMatchInfo> allTemplatesInContext, bool useForListing = false)
+        public TemplateListResolutionResult(IReadOnlyCollection<ITemplateMatchInfo> coreMatchedTemplates)
         {
-            _templateName = templateName;
-            _hasUserInputLanguage = !string.IsNullOrEmpty(userInputLanguage);
             _coreMatchedTemplates = coreMatchedTemplates;
-            _allTemplatesInContext = allTemplatesInContext;
-            _bestTemplateMatchList = null;
         }
-
-        private readonly string _templateName;
-        private readonly bool _hasUserInputLanguage;
 
         private readonly IReadOnlyCollection<ITemplateMatchInfo> _coreMatchedTemplates;
-        private readonly IReadOnlyCollection<ITemplateMatchInfo> _allTemplatesInContext;
+        private IReadOnlyCollection<ITemplateMatchInfo> _exactMatchedTemplates;
+        private IReadOnlyCollection<ITemplateMatchInfo> _partiallyMatchedTemplates;
 
-        public bool TryGetCoreMatchedTemplatesWithDisposition(Func<ITemplateMatchInfo, bool> filter, out IReadOnlyList<ITemplateMatchInfo> matchingTemplates)
+        /// <summary>
+        /// Returns list of exact or partially matched templates by name and exact match by language, filter, baseline (if specified in command paramaters)
+        /// </summary>
+        public IReadOnlyCollection<ITemplateMatchInfo> ExactMatchedTemplates
         {
-            matchingTemplates = _coreMatchedTemplates.Where(filter).ToList();
-            return matchingTemplates.Count != 0;
+            get
+            {
+                if (_exactMatchedTemplates == null)
+                {
+                    // condition is required to filter matches on custom parameters
+                    // TODO: when matching on custom parameters is refactored keep IsMatch condition only
+                    if (_coreMatchedTemplates.Where(t => t.IsInvokableMatch()).Any())
+                    {
+                        _exactMatchedTemplates = _coreMatchedTemplates.Where(t => t.IsInvokableMatch()).ToList();
+                    }
+                    else if (_coreMatchedTemplates.Where(t => t.IsMatch).Any())
+                    {
+                        _exactMatchedTemplates = _coreMatchedTemplates.Where(t => t.IsMatch).ToList();
+                    }
+                    // if there is no match try to match by classification (tag) and other filters ignoring name mismatch.
+                    // in case classification matches given template name in command these templates still to be shown in the list.
+                    // classfication matching is only done in case --list is specified in column, otherwise collection of match disposition doesn't have matchs on classification
+                    // TODO: when matching on tags is implemented via special parameter (--tags) keep IsMatch condition only
+                    else
+                    {
+                        _exactMatchedTemplates = _coreMatchedTemplates.Where(t => t.HasClassificationMatchAndNameMismatch()).ToList();
+                    }
+                }
+                return _exactMatchedTemplates;
+            }
         }
 
-        // If a single template group can be resolved, return it.
-        // If the user input a language, default language results are not considered.
-        // ignoreDefaultLanguageFiltering = true will also cause default language filtering to be ignored. Be careful when using this option.
-        public bool TryGetUnambiguousTemplateGroupToUse(out IReadOnlyList<ITemplateMatchInfo> unambiguousTemplateGroup, bool ignoreDefaultLanguageFiltering = false)
+        /// <summary>
+        /// Returns list of exact or partially matched templates by name and exact match by language, filter, baseline (if specified in command paramaters) grouped by group identity
+        /// </summary>
+        public IReadOnlyCollection<IGrouping<string, ITemplateMatchInfo>> ExactMatchedTemplatesGrouped =>
+            ExactMatchedTemplates.GroupBy(x => x.Info.GroupIdentity, x => !string.IsNullOrEmpty(x.Info.GroupIdentity), StringComparer.OrdinalIgnoreCase).ToList();
+
+        /// <summary>
+        /// Returns list of exact or partially matched templates by name and mismatch in any of the following: language, filter, baseline (if specified in command paramaters)
+        /// </summary>
+        public IReadOnlyCollection<ITemplateMatchInfo> PartiallyMatchedTemplates
         {
-            if (_coreMatchedTemplates.Count == 0)
+            get
             {
-                unambiguousTemplateGroup = null;
-                return false;
-            }
-
-            if (_coreMatchedTemplates.Count == 1)
-            {
-                unambiguousTemplateGroup = new List<ITemplateMatchInfo>(_coreMatchedTemplates);
-                return true;
-            }
-
-            // maybe: only use default language if we're trying to invoke
-            if (!_hasUserInputLanguage && !ignoreDefaultLanguageFiltering)
-            {
-                // only consider default language match dispositions if the user did not specify a language.
-                List<ITemplateMatchInfo> defaultLanguageMatchedTemplates = _coreMatchedTemplates.Where(x =>
-                                                                x.DispositionOfDefaults.Any(y => y.Location == MatchLocation.DefaultLanguage && y.Kind == MatchKind.Exact)
-                                                                && !x.MatchDisposition.Any(z => z.Location == MatchLocation.Context && z.Kind == MatchKind.Mismatch))
-                                                                            .ToList();
-
-                if (TemplateListResolver.AreAllTemplatesSameGroupIdentity(defaultLanguageMatchedTemplates))
+                if (_partiallyMatchedTemplates == null)
                 {
-                    if (defaultLanguageMatchedTemplates.Any(x => !x.HasParameterMismatch() && !x.HasContextMismatch()))
+                    // TODO: clarify if collection should include matching by classification only.
+                    // At the moment defintion of partial match:
+                    // - has exact or partial match by name or short name or classification
+                    // - has mismatch in language, type or baseline
+                    _partiallyMatchedTemplates = _coreMatchedTemplates.Where(t => t.HasNameOrClassificationMatchOrPartialMatch() && t.HasAnyMismatch()).ToList();
+                }
+                return _partiallyMatchedTemplates;
+            }
+        }
+
+        /// <summary>
+        ///  Returns list of exact or partially matched templates by name and mismatch in any of the following: language, filter, baseline (if specified in command paramaters); grouped by group identity
+        /// </summary>
+        public IReadOnlyCollection<IGrouping<string, ITemplateMatchInfo>> PartiallyMatchedTemplatesGrouped =>
+            PartiallyMatchedTemplates.GroupBy(x => x.Info.GroupIdentity, x => !string.IsNullOrEmpty(x.Info.GroupIdentity), StringComparer.OrdinalIgnoreCase).ToList();
+
+
+        /// <summary>
+        /// Returns true when at least one template in unambiguous matches default language
+        /// </summary>
+        public bool HasUnambiguousTemplateGroupForDefaultLanguage => UnambiguousTemplatesForDefaultLanguage.Any();
+
+        /// <summary>
+        /// Returns collecion of templates from unamgibuous group that matches default language
+        /// </summary>
+        public IReadOnlyCollection<ITemplateMatchInfo> UnambiguousTemplatesForDefaultLanguage => UnambiguousTemplateGroup.Where(t => t.HasDefaultLanguageMatch()).ToList();
+
+        /// <summary>
+        /// Returns true when at least one template exactly or partially matched templates by name and exactly matched language, filter, baseline (if specified in command paramaters)
+        /// </summary>
+        public bool HasExactMatches => ExactMatchedTemplates.Any();
+
+        /// <summary>
+        /// Returns true when at least one template exactly or partially matched templates by name but has mismatch in any of the following: language, filter, baseline (if specified in command paramaters)
+        /// </summary>
+        public bool HasPartialMatches => PartiallyMatchedTemplates.Any();
+
+        /// <summary>
+        /// Returns true when at least one template has mismatch in language
+        /// </summary>
+        // as we need to count errors per template group: for language we need to check template groups as templates in single group may have different language
+        // and if one of them can match the filter, then the whole group matches the filter
+        // if all templates in the group has language mismatch, then group has language mismatch
+        public bool HasLanguageMismatch => PartiallyMatchedTemplatesGrouped.Any(g => g.All(t => t.HasLanguageMismatch()));
+
+        /// <summary>
+        /// Returns true when at least one template has mismatch in context (type)
+        /// </summary>
+        public bool HasContextMismatch => PartiallyMatchedTemplates.Any(t => t.HasContextMismatch());
+
+        /// <summary>
+        /// Returns true when at least one template has mismatch in baseline
+        /// </summary>
+        public bool HasBaselineMismatch => PartiallyMatchedTemplates.Any(t => t.HasBaselineMismatch());
+
+        /// <summary>
+        /// Returns true when at least one template has mismatch in author
+        /// </summary>
+        public bool HasAuthorMismatch => PartiallyMatchedTemplates.Any(t => t.HasAuthorMismatch());
+
+        /// <summary>
+        /// Returns true when one and only one template has exact match
+        /// </summary>
+        public bool HasUnambiguousTemplateGroup => ExactMatchedTemplatesGrouped.Count == 1;
+
+        /// <summary>
+        /// Returns list of templates for unambiguous template group, otherwise empty list
+        /// </summary>
+        public IReadOnlyCollection<ITemplateMatchInfo> UnambiguousTemplateGroup => HasUnambiguousTemplateGroup ? ExactMatchedTemplates : new List<ITemplateMatchInfo>();
+
+        /// <summary>
+        /// Returns true if all the templates in unambiguous group have templates in same language
+        /// </summary>
+        public bool AllTemplatesInUnambiguousTemplateGroupAreSameLanguage
+        {
+            get
+            {
+                if (UnambiguousTemplateGroup.Count == 0)
+                {
+                    return false;
+                }
+
+                if (UnambiguousTemplateGroup.Count == 1)
+                {
+                    return true;
+                }
+                    
+                HashSet<string> languagesFound = new HashSet<string>();
+                foreach (ITemplateMatchInfo template in UnambiguousTemplateGroup)
+                {
+                    string language;
+
+                    if (template.Info.Tags != null && template.Info.Tags.TryGetValue("language", out ICacheTag languageTag))
                     {
-                        unambiguousTemplateGroup = defaultLanguageMatchedTemplates.Where(x => !x.HasParameterMismatch() && !x.HasContextMismatch()).ToList();
-                        return true;
+                        language = languageTag.ChoicesAndDescriptions.Keys.FirstOrDefault();
                     }
                     else
                     {
-                        unambiguousTemplateGroup = defaultLanguageMatchedTemplates;
-                        return true;
+                        language = string.Empty;
                     }
-                }
-            }
 
-            List<ITemplateMatchInfo> paramFiltered = _coreMatchedTemplates.Where(x => !x.HasParameterMismatch() && !x.HasContextMismatch()).ToList();
-            if (TemplateListResolver.AreAllTemplatesSameGroupIdentity(paramFiltered))
-            {
-                unambiguousTemplateGroup = paramFiltered;
-                return true;
-            }
-
-            if (TemplateListResolver.AreAllTemplatesSameGroupIdentity(_coreMatchedTemplates))
-            {
-                unambiguousTemplateGroup = new List<ITemplateMatchInfo>(_coreMatchedTemplates);
-                return true;
-            }
-
-            unambiguousTemplateGroup = null;
-            return false;
-        }
-
-        // Note: This method does not consider default language matches / mismatches
-        public bool TryGetAllInvokableTemplates(out IReadOnlyList<ITemplateMatchInfo> invokableTemplates)
-        {
-            IEnumerable<ITemplateMatchInfo> invokableMatches = _coreMatchedTemplates.Where(x => x.IsInvokableMatch());
-
-            if (invokableMatches.Any())
-            {
-                invokableTemplates = invokableMatches.ToList();
-                return true;
-            }
-
-            invokableTemplates = null;
-            return false;
-        }
-
-        public enum SingularInvokableMatchCheckStatus
-        {
-            None,
-            NoMatch,
-            SingleMatch,
-            AmbiguousChoice,
-            AmbiguousPrecedence
-        }
-
-        public bool TryGetSingularInvokableMatch(out ITemplateMatchInfo template, out SingularInvokableMatchCheckStatus resultStatus)
-        {
-            IReadOnlyList<ITemplateMatchInfo> invokableMatches = _coreMatchedTemplates.Where(x => x.IsInvokableMatch()).ToList();
-            IReadOnlyList<ITemplateMatchInfo> languageFilteredInvokableMatches;
-
-            if (_hasUserInputLanguage)
-            {
-                languageFilteredInvokableMatches = invokableMatches;
-            }
-            else
-            {
-                // check for templates with the default language
-                languageFilteredInvokableMatches = invokableMatches.Where(x => x.DispositionOfDefaults.Any(y => y.Location == MatchLocation.DefaultLanguage && y.Kind == MatchKind.Exact)).ToList();
-
-                // no candidate templates matched the default language, continue with the original candidates.
-                if (languageFilteredInvokableMatches.Count == 0)
-                {
-                    languageFilteredInvokableMatches = invokableMatches;
-                }
-            }
-
-            if (languageFilteredInvokableMatches.Count == 1)
-            {
-                template = languageFilteredInvokableMatches[0];
-                resultStatus = SingularInvokableMatchCheckStatus.SingleMatch;
-                return true;
-            }
-
-            // if multiple templates in the group have single starts with matches on the same parameter, it's ambiguous.
-            // For the case where one template has single starts with, and another has ambiguous - on the same param:
-            //      The one with single starts with is chosen as invokable because if the template with an ambiguous match
-            //      was not installed, the one with the singluar invokable would be chosen.
-            HashSet<string> singleStartsWithParamNames = new HashSet<string>();
-            foreach (ITemplateMatchInfo checkTemplate in languageFilteredInvokableMatches)
-            {
-                IList<string> singleStartParamNames = checkTemplate.MatchDisposition.Where(x => x.Location == MatchLocation.OtherParameter && x.Kind == MatchKind.SingleStartsWith).Select(x => x.InputParameterName).ToList();
-                foreach (string paramName in singleStartParamNames)
-                {
-                    if (!singleStartsWithParamNames.Add(paramName))
+                    if (!string.IsNullOrEmpty(language))
                     {
-                        template = null;
-                        resultStatus = SingularInvokableMatchCheckStatus.AmbiguousChoice;
+                        languagesFound.Add(language);
+                    }
+
+                    if (languagesFound.Count > 1)
+                    {
                         return false;
                     }
                 }
-            }
-
-            ITemplateMatchInfo highestInGroupIfSingleGroup = TemplateListResolver.FindHighestPrecedenceTemplateIfAllSameGroupIdentity(languageFilteredInvokableMatches, out bool ambiguousGroupIdResult);
-
-            if (highestInGroupIfSingleGroup != null)
-            {
-                template = highestInGroupIfSingleGroup;
-                resultStatus = SingularInvokableMatchCheckStatus.SingleMatch;
                 return true;
-            }
-            else if (ambiguousGroupIdResult)
-            {
-                template = null;
-                resultStatus = SingularInvokableMatchCheckStatus.AmbiguousPrecedence;
-                return false;
-            }
-
-            template = null;
-            resultStatus = SingularInvokableMatchCheckStatus.NoMatch;
-            return false;
-        }
-
-        private IReadOnlyList<ITemplateMatchInfo> _bestTemplateMatchList;
-        private IReadOnlyList<ITemplateMatchInfo> _bestTemplateMatchListIgnoringDefaultLanguageFiltering;
-
-        public IReadOnlyList<ITemplateMatchInfo> GetBestTemplateMatchList(bool ignoreDefaultLanguageFiltering = false)
-        {
-            if (ignoreDefaultLanguageFiltering)
-            {
-                if (_bestTemplateMatchList == null)
-                {
-                    _bestTemplateMatchList = BaseGetBestTemplateMatchList(ignoreDefaultLanguageFiltering);
-                }
-
-                return _bestTemplateMatchList;
-            }
-            else
-            {
-                if (_bestTemplateMatchListIgnoringDefaultLanguageFiltering == null)
-                {
-                    _bestTemplateMatchListIgnoringDefaultLanguageFiltering = BaseGetBestTemplateMatchList(ignoreDefaultLanguageFiltering);
-                }
-
-                return _bestTemplateMatchListIgnoringDefaultLanguageFiltering;
-            }
-        }
-
-        // The core matched templates should not need additioanl default language filtering.
-        // The default language dispositions are stored in a different place than the other dispositions,
-        // and are not considered for most match filtering.
-        private IReadOnlyList<ITemplateMatchInfo> BaseGetBestTemplateMatchList(bool ignoreDefaultLanguageFiltering)
-        {
-            IReadOnlyList<ITemplateMatchInfo> templateList;
-            if (TryGetUnambiguousTemplateGroupToUse(out templateList, ignoreDefaultLanguageFiltering))
-            {
-                return templateList;
-            }
-            else if (!string.IsNullOrEmpty(_templateName) && TryGetAllInvokableTemplates(out templateList))
-            {
-                return templateList;
-            }
-            else if (TryGetCoreMatchedTemplatesWithDisposition(x => x.IsMatch, out templateList))
-            {
-                return templateList;
-            }
-            else if (TryGetCoreMatchedTemplatesWithDisposition(x => x.IsMatchExceptContext(), out templateList))
-            {
-                return templateList;
-            }
-            else if (TryGetCoreMatchedTemplatesWithDisposition(x => x.IsPartialMatch, out templateList))
-            {
-                return templateList;
-            }
-            else if (TryGetCoreMatchedTemplatesWithDisposition(x => x.IsPartialMatchExceptContext(), out templateList))
-            {
-                return templateList;
-            }
-            else
-            {
-                templateList = _allTemplatesInContext.ToList();
-                return templateList;
             }
         }
